# api/docs/swagger.yaml
basePath: /v1
definitions:
...
  models.AllProducts:
    properties:
      products:
        items:
          $ref: '#/definitions/models.ProductRes'
        type: array
    type: object
...
  models.FailureInfo:
    properties:
      code:
        type: integer
      error: {}
      message:
        type: string
    type: object
...
  models.MediaLink:
    properties:
      mediaLink:
        type: string
    type: object
...
  models.ProductReq:
    properties:
      brand_id:
        type: integer
      camera:
        type: string
      category_id:
        type: integer
      characteristics:
        additionalProperties:
          type: string
        example:
          key: value
        type: object
      description:
        type: string
      dioganal:
        type: number
      display_type:
        type: string
      media_link:
        items:
          $ref: '#/definitions/models.MediaLink'
        type: array
      os_type:
        type: string
      price:
        type: number
      title:
        type: string
    type: object
  models.ProductRes:
    properties:
      brand_id:
        type: integer
      camera:
        type: string
      category_id:
        type: integer
      characteristics:
        additionalProperties:
          type: string
        example:
          key: value
        type: object
      created_at:
        type: string
      description:
        type: string
      dioganal:
        type: number
      display_type:
        type: string
      id:
        type: integer
      media_link:
        items:
          $ref: '#/definitions/models.MediaLink'
        type: array
      os_type:
        type: string
      price:
        type: number
      title:
        type: string
      updated_at:
        type: string
    type: object
  models.SuccessInfo:
    properties:
      message:
        type: string
      status_code:
        type: integer
    type: object
  models.UpdateMediaLink:
    properties:
      id:
        type: integer
      media_link:
        type: string
    type: object
...
  models.UpdateProductReq:
    properties:
      camera:
        type: string
      characteristics:
        additionalProperties:
          type: string
        example:
          key: value
        type: object
      description:
        type: string
      dioganal:
        type: number
      display_type:
        type: string
      id:
        type: integer
      media_link:
        items:
          $ref: '#/definitions/models.UpdateMediaLink'
        type: array
      os_type:
        type: string
      price:
        type: number
      title:
        type: string
    type: object
  models.UpdateUserReq:
    properties:
      id:
        type: integer
      phone_number:
        type: string
    type: object
  models.UserReq:
    properties:
      phone_number:
        example: "998"
        type: string
    type: object
  models.UserRes:
    properties:
      created_at:
        type: string
      id:
        type: integer
      phone_number:
        type: string
      updated_at:
        type: string
    type: object
...
  repo.MediaRes:
    properties:
      id:
        type: integer
      mediaLink:
        type: string
      productId:
        type: integer
    type: object
...
  repo.ProductResponse:
    properties:
      brandId:
        type: integer
      camera:
        type: string
      categoryId:
        type: integer
      characteristics:
        type: string
      createdAt:
        type: string
      description:
        type: string
      dioganal:
        type: number
      displayType:
        type: string
      id:
        type: integer
      mediaLinks:
        items:
          $ref: '#/definitions/repo.MediaRes'
        type: array
      osType:
        type: string
      price:
        type: number
      rating:
        type: number
      title:
        type: string
      updatedAt:
        type: string
    type: object
host: 100.26.188.92:7070
info:
  contact:
    email: gofurovmurtazoxon@gmail.com
    name: Murtazoxon
    url: https://t.me/murtazokhon_gofurov
  description: This is climate_uz server api server
  termsOfService: 2 term climate_uz
  title: Climate_uz api
  version: "1.0"
paths:
...
  /product:
    post:
      consumes:
      - application/json
      description: Through this api can create product
      parameters:
      - description: ProductRequest
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/models.ProductReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/repo.ProductResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureInfo'
      security:
      - BearerAuth: []
      summary: create new product
      tags:
      - Product
    put:
      consumes:
      - application/json
      description: this method edit product info
      parameters:
      - description: UpdateProductRequest
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/models.UpdateProductReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repo.ProductResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureInfo'
      security:
      - BearerAuth: []
      summary: edit product info
      tags:
      - Product
  /product/{id}:
    delete:
      consumes:
      - application/json
      description: this method delete product info by productId
      parameters:
      - description: product_id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessInfo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureInfo'
      security:
      - BearerAuth: []
      summary: delete product info
      tags:
      - Product
    get:
      consumes:
      - application/json
      description: this method get product by Id
      parameters:
      - description: product_id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repo.ProductResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureInfo'
      summary: get one product
      tags:
      - Product
  /products:
    get:
      consumes:
      - application/json
      description: this method get all products list
      parameters:
      - description: page
        in: query
        name: page
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AllProducts'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureInfo'
      summary: get all products
      tags:
      - Product
  /user:
    patch:
      consumes:
      - application/json
      description: This method update user info
      parameters:
      - description: UpdateUserReq
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UpdateUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureInfo'
      security:
      - BearerAuth: []
      summary: Update user info
      tags:
      - User
    post:
      consumes:
      - application/json
      description: This function create new user
      parameters:
      - description: UserRequest
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UserReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.UserRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureInfo'
      summary: add user phone_number
      tags:
      - User
  /user/{id}:
    delete:
      consumes:
      - application/json
      description: This method delete user info
      parameters:
      - description: user_id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureInfo'
      security:
      - BearerAuth: []
      summary: Delete user info
      tags:
      - User
    get:
      consumes:
      - application/json
      description: This function get user by id
      parameters:
      - description: user_id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureInfo'
      security:
      - BearerAuth: []
      summary: Get user
      tags:
      - User
  /users:
    get:
      consumes:
      - application/json
      description: This method is to get all users
      parameters:
      - description: page
        in: query
        name: page
        required: true
        type: string
      - description: limit
        in: query
        name: limit
        required: true
        type: integer
      - description: search
        in: query
        name: search
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureInfo'
      summary: Get all users
      tags:
      - User
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"